requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
Werkzeug==1.0.1
flask-cors==3.0.10
simplejson==3.16.0
requests_toolbelt
Flask-RESTful==0.3.8
pymysql==1.0.2
python-dateutil==2.8.1
bcrypt==3.2.0
jwcrypto==0.8
PyJWT==1.7.1
setuptools==45.2.0 
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
Pillow==7.0.0
flask_session
python-dotenv
Flask-Caching
flask_mail